fix(UI): Limit grid size to 1-100, default 16
When the user is prompted for the new grid size, they are prompted to
enter an integer between 1 and 100. Previously, we did not have any
handling for the case where the user entered an integer outside of that
range. Now, if the user enters a number outside of that range, generate
a grid of the default size, 16.